Work in progress. The UltimateII app builds again, but is not completely functional yet. Still many little things to fix, using some new infrastructure. Hey, at least the configuration menu is working, and the SdCard is showing in the main menu. Yey! DO NOT USE!

+++ software/io/c64/keyboard_c64.h
@@ -30,7 +30,7 @@ public:
     
     void scan(void);
     void set_delays(int, int);
-    uint8_t getch(void);
+    int  getch(void);
     void wait_free(void);
     void clear_buffer(void);
 };
